+++ src/i_video.h
@@ -93,7 +93,8 @@ void I_DeferredInitColor(void);
 // [Nugget] -----------------------------------------------------------------/
 
 extern int gamma2;
-byte I_GetNearestColor(byte *palette, int r, int g, int b);
+byte I_GetNearestColor(const byte *palette, int r, int g, int b);
+byte I_GetNearestColorLinear(const byte *palette, int red, int green, int blue);
 
 boolean I_WritePNGfile(char *filename); // [FG] screenshots in PNG format