+++ src/mn_menu.h
@@ -115,7 +115,6 @@ int MN_GetPixelWidth(const char *ch);
 void MN_DrawString(int cx, int cy, int color, const char *ch);
 
 void M_StartSound(int sound_id);
-void M_StartSoundOptional(const int opt_sound_id, const int sound_id); // [Nugget]
 
 void M_StartCustomSkill(const int mode); // [Nugget] Custom Skill