+++ src/sounds.c
@@ -258,6 +258,21 @@ sfxinfo_t original_S_sfx[NUMSFX] = {
   SOUND("plosml", sg_oof,    96),
   SOUND("lavsml", sg_oof,    96),
 
+  // optional menu and intermission sounds
+  SOUND("mnuopn", sg_none, 60),
+  SOUND("mnucls", sg_none, 60),
+  SOUND("mnuact", sg_none, 60),
+  SOUND("mnubak", sg_none, 60),
+  SOUND("mnumov", sg_none, 60),
+  SOUND("mnusli", sg_none, 60),
+  SOUND("mnusel", sg_none, 60),
+  SOUND("mnuerr", sg_none, 60),
+  SOUND("inttic", sg_none, 60),
+  SOUND("inttot", sg_none, 60),
+  SOUND("intnex", sg_none, 60),
+  SOUND("intnet", sg_none, 60),
+  SOUND("intdms", sg_none, 60),
+
   // [Nugget] /---------------------------------------------------------------
 
   // [NS] New optional sounds.
@@ -266,19 +281,6 @@ sfxinfo_t original_S_sfx[NUMSFX] = {
   SOUND("locked", sg_oof,    78),
   SOUND("keyup",  sg_itemup, 78),
   SOUND("keybnk", sg_oof,    78), // [Nugget]
-  // [NS] Optional menu/intermission sounds.
-  SOUND("mnuopn", sg_none,   60),
-  SOUND("mnucls", sg_none,   60),
-  SOUND("mnuact", sg_none,   60),
-  SOUND("mnubak", sg_none,   60),
-  SOUND("mnumov", sg_none,   60),
-  SOUND("mnusli", sg_none,   60),
-  SOUND("mnuerr", sg_none,   60),
-  SOUND("inttic", sg_none,   60),
-  SOUND("inttot", sg_none,   60),
-  SOUND("intnex", sg_none,   60),
-  SOUND("intnet", sg_none,   60),
-  SOUND("intdms", sg_none,   60),
 
   SOUND("ppai25", sg_none, 96),
   SOUND("ppai50", sg_none, 96),