+++ CHANGELOG.md
@@ -4,8 +4,10 @@ None.
 
 ## Changes
 
+- **Merged changes from Woof! post-16.0.0**
 - **Made the NUGHUD have its own slot**
 - **Made automap marks flash brighter when using _Highlight P.O.I.'s_**
+- **Improved low-resolution post-process**
 
 ## Bug Fixes